refactor: fabric functions
